Rename db.models.ModelGenerator
